libutil: factor out common util funcs
liblsd comes from an internal LLNL called lsd-tools,
so keep that source separate from other utility functions
specific to this project.

Relocate hpcrc.[ch] and util.[ch] to libutil.

Build all libraries with libtool to avoid mixing pic
and non-pic code when creating shared libraries.

Change AP_CPPFLAGS to only -I$(top_srcdir), and then
have each .c file tht uses internal library functions include
headers via a path relative to top_srcdir.
