+++ Makefile.am
@@ -1,1 +1,1 @@
-SUBDIRS = libvxi11 libics liblsd src man etc test
+SUBDIRS = libvxi11 libics liblsd libutil src man etc test
