+++ libutil/Makefile.am
@@ -1,0 +1,13 @@
+AM_CFLAGS = @GCCWARN@
+
+noinst_LTLIBRARIES = libutil.la
+
+libutil_la_SOURCES = \
+	argv.c \
+	argv.h \
+	hprintf.c \
+	hprintf.h \
+	util.c \
+	util.h \
+	hpcrc.c \
+	hpcrc.h
