+++ libutil/hpcrc.c
