+++ libutil/hpcrc.h
