+++ libutil/hprintf.c
