+++ libutil/hprintf.h
