+++ libutil/util.c
