+++ libutil/util.h
