+++ src/hp6038.c
@@ -35,7 +35,7 @@
 #include <math.h>
 #include <stdint.h>
 
-#include "util.h"
+#include "libutil/util.h"
 #include "gpib.h"
 
 #define INSTRUMENT "hp6038"
